backend/requirements.txt
amqp==5.3.1
asgiref==3.9.0
billiard==4.2.1
celery==5.5.3
cffi==2.0.0
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
dj-database-url==3.0.1
Django==5.2.4
django-cors-headers==4.7.0
django-environ==0.12.0
django-extensions==4.1
django-filter==25.1
django-nested-admin==4.1.6
django-taggit==6.1.0
django_celery_results==2.6.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
drf-nested-routers==0.94.2
et_xmlfile==2.0.0
jsonpointer==2.1
kombu==5.5.4
numpy==2.3.2
openpyxl==3.1.5
openslide-bin==4.0.0.8
openslide-python==1.4.2
packaging==25.0
pandas==2.3.1
pillow==11.3.0
prompt_toolkit==3.0.52
psycopg2==2.9.11
psycopg2-binary==2.9.11
pycparser==2.23
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-monkey-business==1.1.0
pytz==2025.2
pyvips==3.0.0
redis==6.4.0
rest-framework-simplejwt==0.0.2
setuptools==78.1.1
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
wheel==0.45.1
whitenoise==6.11.0
